requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decouple==0.0.7
Django==5.1.2
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
pytest-mock==3.14.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.4.0
